+++ src/Nethereum.RPC/Eth/Transactions/IEthFeeHistory.cs
@@ -15,7 +15,7 @@ namespace Nethereum.RPC.Eth.Transactions
         /// <param name="rewardPercentiles">A monotonically increasing list of percentile values to sample from each block's effective priority fees per gas in ascending order, weighted by gas used.
         /// Floating point value between 0 and 100.</param>
         /// <returns></returns>
-        Task<FeeHistoryResult> SendRequestAsync(HexBigInteger blockCount, BlockParameter highestBlockNumber, double[] rewardPercentiles = null, object id = null);
+        Task<FeeHistoryResult> SendRequestAsync(HexBigInteger blockCount, BlockParameter highestBlockNumber, decimal[] rewardPercentiles = null, object id = null);
 
         /// <summary>
         /// Builds the Request, to return base fee per gas and transaction effective priority fee per gas history for the requested block range if available. The range between headBlock-4 and headBlock is guaranteed to be available while retrieving data from the pending block and older history are optional to support. For pre-EIP-1559 blocks the gas prices are returned as rewards and zeroes are returned for the base fee per gas.
@@ -25,6 +25,6 @@ namespace Nethereum.RPC.Eth.Transactions
         /// <param name="rewardPercentiles">A monotonically increasing list of percentile values to sample from each block's effective priority fees per gas in ascending order, weighted by gas used.
         /// Floating point value between 0 and 100.</param>
         /// <returns></returns>
-        RpcRequest BuildRequest(HexBigInteger blockCount, BlockParameter highestBlockNumber, double[] rewardPercentiles = null, object id = null);
+        RpcRequest BuildRequest(HexBigInteger blockCount, BlockParameter highestBlockNumber, decimal[] rewardPercentiles = null, object id = null);
     }
 }
